+++ components/upload/style/index.jsx
@@ -1,0 +1,2 @@
+import '../../../style/index.less';
+import './index.less';
