+++ components/upload/uploadList.jsx
@@ -131,9 +131,9 @@ export default class UploadList extends React.Component {
                   >
                     <Icon type='eye-o' />
                   </a>
-                  <Icon type='close' onClick={() => this.handleClose(file)} />
+                  <Icon type='delete' onClick={() => this.handleClose(file)} />
                 </span>
-              ) : <Icon type='close-circle' onClick={() => this.handleClose(file)} />
+              ) : <Icon type='close' onClick={() => this.handleClose(file)} />
             }
           </div>
           {progress}
